requirements.txt
Flask==2.0.1
tensorflow-cpu==2.5.0
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
gunicorn==20.1.0
pandas==1.1.5